+++ praw/objector.py
@@ -72,33 +72,18 @@ class Objector:
                 if "conversation" in data
                 else data.pop("conversations")
             )
-            parser: "praw.models.ModmailConversation" = self.parsers[
-                "ModmailConversation"
-            ]
+            parser = self.parsers["ModmailConversation"]
             parser._convert_conversation_objects(data, self._reddit)
-        elif {"messages", "modActions"}.issubset(data):
+        elif {"messages", "modActions"}.issubset(data) or {
+            "legacyFirstMessageId",
+            "state",
+        }.issubset(data):
             parser = self.parsers["ModmailConversation"]
         elif {"conversationIds", "conversations", "messages"}.issubset(data):
-            conversations = []
-            temp_data = {
-                thing_type: data.pop(thing_type, {})
-                for thing_type in ["messages", "modActions"]
-            }
-            for conversation_id in data["conversationIds"]:
-                conversation_data = {
-                    **data["conversations"][conversation_id],
-                    "messages": [],
-                    "modActions": [],
-                }
-                for object_id in conversation_data["objIds"]:
-                    conversation_data[object_id["key"]].append(
-                        self._objectify_dict(
-                            temp_data[object_id["key"]].pop(object_id["id"])
-                        )
-                    )
-                conversations.append(conversation_data)
-            data["conversations"] = conversations
-            del temp_data
+            data["conversations"] = [
+                data["conversations"][conversation_id]
+                for conversation_id in data["conversationIds"]
+            ]
             data = snake_case_keys(data)
             parser = self.parsers["ModmailConversations-list"]
         elif {"actionTypeId", "author", "date"}.issubset(data):